vod-toolkits/playauth-generate/Python/requirements.txt
PyJWT>=2.0.0
